feat: support for internal high cycle area flash (#449)
* Allow enable NonMaskableInt_IRQn interrupt

* Fix Flash Write and cleanup

Before this change, writing data to flash for STM32WBA/STM32H5 only considered only the first 16 bytes.
The remaining data was not processed. Fix this by iterating over an fitting MemoryRange like it is done
for the other MCU families.
This change also cleans up writing flash method calls for the different families.

* Add FlashInternalHighCycleAreaStm

* Increase flash latency by 2

Needed for the use of OTP, read-only and flash high-cycle data.
From datasheet:
"During a read access, two wait states are added in addition to the memory
wait states. These wait states are necessary to parse the data buffer."