feat: improve Pinout Table Generation (#618)
* Allow setting the PinConfigTypeStm type

When provided, take the Pin's `type` attribute as value for PinConfigTypeStm, otherwise use the concatenation of Peripheral name and Pin name, as it was before this change.

* Allow providing custom PinoutTableItems.xml

The PinoutTableItems.xml defines hardware specific IO configuration for peripherals like Speed or WeakPull. This change allows providing a xml file customized for individual hardware.